+++ src/pretalx/schedule/models/schedule.py
@@ -364,6 +364,28 @@ class Schedule(LogMixin, models.Model):
                         ),
                     }
                 )
+        if talk.room:
+            overlaps = (
+                TalkSlot.objects.filter(
+                    schedule=self, room=talk.room
+                )
+                .filter(
+                    models.Q(start__lt=talk.start, end__gt=talk.start)
+                    | models.Q(start__lt=talk.real_end, end__gt=talk.real_end)
+                    | models.Q(start__gt=talk.start, end__lt=talk.real_end)
+                ).exclude(pk=talk.pk)
+                .exists()
+            )
+            if overlaps:
+                warnings.append(
+                    {
+                        "type": "room_overlap",
+                        "message": _(
+                            "There's an overlapping session scheduled in this room."
+                        ),
+                    }
+                )
+
         for speaker in talk.submission.speakers.all():
             if with_speakers:
                 profile = speaker.event_profile(event=self.event)